+++ packages/libs/ui/src/components/icons/LyIcon.vue
@@ -1,12 +1,13 @@
 <script lang="ts" setup>
-import { IconName, getIconByName } from './Icons';
-import { computed } from 'vue';
 import { IconOptionsIF } from '@/types';
-import { includesUtilityClass } from '@/helpers';
-import { t, Translatable } from '@/i18n';
+import { Translatable } from '@/i18n';
+import {getIconLibrary} from "@/components/icons/registries";
+import {computed} from "vue";
 
 export interface IProps {
-  name?: IconName;
+  name?: string;
+  lib?: string;
+  libBindings?: any;
   title?: Translatable;
   options?: IconOptionsIF;
   scaleTo?: number;
@@ -15,59 +16,21 @@ export interface IProps {
 
 const props = withDefaults(defineProps<IProps>(), {
   name: '',
+  lib: 'ly',
+  libBindings: undefined,
   options: undefined,
   title: undefined,
   scaleTo: 0,
   autoScale: false,
 });
 
-const name = computed<IconName>(() => props.name || props.options?.name || '');
-const definition = computed(() => getIconByName(name.value));
-const styleObject = computed(() => (props.options?.fill ? { fill: props.options.fill } : {}));
-
-function getClassNames(attrClasses: any) {
-  const result = {
-    icon: true,
-    'w-4': !includesUtilityClass(attrClasses, 'w'),
-    inline: true,
-    [`icon-${name.value}`]: true,
-  };
-
-  if (props.options?.color) {
-    result[props.options.color.toString()] = true;
-  }
-
-  return result;
-}
-
-const transform = computed<string | undefined>(() => {
-  if (!props.scaleTo && !props.autoScale) return;
-
-  const scaleTo = props.scaleTo || 24;
-
-  try {
-    const size = parseInt(definition.value?.viewBox.split(' ').at(-1) as string);
-    if (size === scaleTo) return;
-    const ratio = scaleTo / size;
-    return `scale(${ratio})`;
-  } catch (e) {
-    console.log(e);
-  }
-
-  return;
-});
+const library = computed(() => getIconLibrary(props.lib));
+const iconComponent = computed(() => library.value?.component);
+const iconBindings = computed(() => library.value?.getBindings(props));
 </script>
 
 <template>
-  <svg
-    v-if="definition"
-    aria-hidden="true"
-    :class="getClassNames($attrs.class)"
-    :style="styleObject"
-    :viewBox="definition.viewBox">
-    <title v-if="title">{{ t(title) }}</title>
-    <path v-for="path in definition.paths" :key="path" :d="path" :transform="transform"></path>
-  </svg>
+  <component v-if="iconComponent" :is="iconComponent" v-bind="iconBindings" />
 </template>
 
 <style>