+++ packages/libs/ui/src/components/icons/registries/icon-library.registry.ts
@@ -1,0 +1,11 @@
+import {IconLibraryIF} from "@/components/icons/interfaces/icon-library.interface";
+
+const libraries = new Map<string, IconLibraryIF>();
+
+export function registerIconLibrary(lib: IconLibraryIF) {
+  libraries.set(lib.id, lib);
+}
+
+export function getIconLibrary(id: string) {
+  return libraries.get(id);
+}