+++ packages/libs/ui/src/components/icons/registries/index.ts
@@ -1,0 +1,2 @@
+export * from './ly-svg-icon.registry';
+export * from './icon-library.registry';